Move SignalMapper classes from net.maswag.falcaun.parser to net.maswag.falcaun
