+++ core/src/main/java/net/maswag/falcaun/package-info.java
@@ -8,7 +8,7 @@
  * to enable effective falsification of complex systems.</p>
  *
  * @see net.maswag.falcaun.parser.AbstractTemporalLogic
- * @see net.maswag.falcaun.parser.SignalMapper
+ * @see net.maswag.falcaun.SignalMapper
  * @see net.maswag.falcaun.ContinuousNumericSUL
  */
 package net.maswag.falcaun;
