+++ core/src/test/java/net/maswag/falcaun/SimpleSignalMapperTest.java
@@ -1,10 +1,11 @@
-package net.maswag.falcaun.parser;
+package net.maswag.falcaun;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import net.maswag.falcaun.IOSignalPiece;
-import net.maswag.falcaun.parser.SimpleSignalMapper;
+import net.maswag.falcaun.SignalMapper;
+import net.maswag.falcaun.SimpleSignalMapper;
 
 import java.io.IOException;
 import java.util.ArrayList;
