+++ matlab/src/test/java/net/maswag/falcaun/simulink/SimulinkSULVerifierTest.java
@@ -7,12 +7,12 @@ import net.maswag.falcaun.AdaptiveSTLUpdater;
 import net.maswag.falcaun.ArgParser;
 import net.maswag.falcaun.IOSignalPiece;
 import net.maswag.falcaun.NumericSULMapper;
+import net.maswag.falcaun.SimpleSignalMapper;
 import net.maswag.falcaun.StaticSTLList;
 import net.maswag.falcaun.StopDisprovedEQOracle;
 import net.maswag.falcaun.ArgParser.GASelectionKind;
 import net.maswag.falcaun.StopDisprovedEQOracle.StaticLTLList;
 import net.maswag.falcaun.parser.STLFactory;
-import net.maswag.falcaun.parser.SimpleSignalMapper;
 import net.maswag.falcaun.parser.TemporalLogic;
 import net.maswag.falcaun.simulink.SimulinkSULVerifier;
